sayso/backend/requirements.txt
fastapi
uvicorn
openai-whisper
python-multipart
openai
python-dotenv 